So many changes, may not use in the future

+++ pages/products/delete/[...id].js
@@ -1,28 +1,36 @@
-import Layout from "@/components/Layout";
+import { useEffect, useState, useContext } from "react";
 import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
 import axios from "axios";
+import Layout from "@/components/Layout";
+import {DemoContext} from "@/contexts/demoContext";  // assuming you've stored it in contexts folder
 
 // Using the id of the product to know the products name for later in the page
 export default function DeleteProductPage() {
   const router = useRouter();
+  const demoData = useContext(DemoContext); // Get demo data from context
   const [productInfo, setProductInfo] = useState();
   const { id } = router.query;
 
   useEffect(() => {
     if (!id) {
+      setProductInfo(demoData); // Set demo data if no ID is provided
       return;
     }
     axios.get("/api/products?id=" + id).then((response) => {
       setProductInfo(response.data);
     });
-  }, [id]);
+  }, [id, demoData]);
 
   function goBack() {
     router.push("/products");
   }
 
   async function deleteProduct() {
+    if(productInfo === demoData) {
+      alert("This is demo data. No need to delete!");
+      goBack();
+      return;
+    }
     await axios.delete("/api/products?id=" + id);
     goBack();
   }
@@ -43,3 +51,6 @@ export default function DeleteProductPage() {
     </Layout>
   );
 }
+
+
+
